Added ability to instrument when exceptions happen
Occasionally, the instrumented code will raise an exception, and we
would still want the subscribers to be notified. For example, if an HTTP
client fails by raising an exception, a subscriber doing logging should
still be able to log the failed request.

```
instrument("example.request") do
  client.post("http://url.that.fails.example/")
end
```

This will call the subscriber with an event that contains the exception
within the payload under a key `:exception`. This implementation also
re-raises the original error, so the callstack is unchanged.